Zoom crossword
A zoom that would fit the crossword in the available space is calculated
on each resize.  (This is a bit wonky right now.)  If it is greater than
one, the zoom is fixed at that level.  If it is less than one, the zoom
is allowed to vary between that and one.  Ideally Ctrl+Scroll would
adjust it, but that's not working at the moment.  Instead, ,/. do the
zooming.

The grid is scrolled to reveal the selected word, if necessary and if it
fits.  If it doesn't fit, the current cell is centered.