Tweak handling of cursor keys
When cursor keys change the selected direction, they will not take a
step unless the grid cell is filled in.  In my use, at least, this seems
to be the behavior I expect.  While filling things in, I want to be able
to switch directions easily; when most things are filled out, I want to
move around the grid easily.