Fix toolbar on device
For some reason, touching toolbar elements only generates touchstart and
touchend events, and not click events.  Luckily, UI.button(id).click()
abstracts this away and gets things working both on the desktop and on
the device.  I thought I had tried this before without it working, but
it seems to be good now.

Ref. #9.