Kludge to fix back button in toolbar
As noted in the previous commit, there are no click events for the
toolbar on the device.  But the SDK only assigns a click listener for
the back button.  Here, we add the robust handler via UI.button.  Note
that this will fire after the build in one (if it's triggered), so we'll
end up popping two pages off the stack if we ever have three or more
pages on it.  See LP #1302284.

Closes #9.