Add scrollbars for puzzle
These are just to show you where you are; you can't interact with them.
Note that animation is now indicated on the #grid; the various things
that are animated look to this to decide what to do.  Because of this,
we've taken out the event handler to remove the animation class.  We
didn't really need that anyway.