Fix grid layout
Two issues: First we need to set the border-spacing now, possibly
because we get a tbody element.  Second, the rows are set to have
white-space: pre, so we need to avoid extraneous white space around our
tags.  To keep the formatting nice in the code, we comment out all the
whitespace we don't want.