Support opening PUZ files in Android
Also, go to the beginning of the history stack before loading files from
outside the app.  Note that this can't be done in a loop; we have to
wait for a callback after moving back in history to check the new
location.

We are making using of the WebIntents plugin
(https://github.com/Initsogar/cordova-webintent.git), but this won't be
automatically set up if this repository is cloned.  Also, we've edited
platforms/android/AndroidManifest.xml directly, but we're not version
controlling that file.